fixed ProductsPage, ProductsTest, xml
