+++ negativeTestFF.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
 <suite name="Smoke Tests" parallel="classes" thread-count="2">
-    <parameter name="browser" value="chrome"/>
-    <test name="Smoke tests">
+    <parameter name="browser" value="firefox"/>
+    <test name="Negative tests">
         <groups>
             <run>
-                <include name="smoke"/>
+                <include name="negative"/>
             </run>
         </groups>
         <packages>
