+++ smokeTestFF.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
 <suite name="Smoke Tests" parallel="classes" thread-count="2">
-    <parameter name="browser" value="chrome"/>
+    <parameter name="browser" value="firefox"/>
     <test name="Smoke tests">
         <groups>
             <run>
